fix(menu): address review - safer OpenXR teardown, drop the menu refactor
Cross-engine review findings:
- request the OpenXR exit before xrWaitFrame instead of after it, so the
  request-failure break can never strand a waited frame with no matching
  xrBeginFrame/xrEndFrame.
- release the XR objects explicitly before `process::exit` (which runs no
  destructors), and end the post-finish() wait as soon as the Android pump
  reports Destroy rather than always burning the full second.
- revert the `effects_for_action` extraction and its unit test: the mapping
  it asserted is defined one screen above it and already covered by the
  existing click test plus the new e2e.
- hoist the duplicated menu/VR-panel test geometry into
  test/helpers/frontend-menu.ts, shared with menu-audio.
