+++ tools/shock2-sdk/test/menu-quit.e2e.test.ts
@@ -1,7 +1,8 @@
 import assert from "node:assert/strict";
 import { test } from "node:test";
 
-import { GameServer, PLAYER_EYE_HEIGHT_WORLD } from "../src/index.js";
+import { GameServer } from "../src/index.js";
+import { AIM_AT_PANEL, menuEntry, panelPoint } from "./helpers/frontend-menu.js";
 
 // End-to-end test for the main menu's Quit entry: clicking it must reach
 // `Game::should_quit`, which is what every runtime shuts down on (the desktop
@@ -15,11 +16,8 @@ import { GameServer, PLAYER_EYE_HEIGHT_WORLD } from "../src/index.js";
 // `undefined` and both assertions below fail.
 const e2eEnabled = process.env.SHOCK2_E2E === "1";
 
-// MAINR.BIN stacks six 179x60 buttons at x=400 on a 76px pitch, on the 640x480
-// UI canvas; Quit is the last one (index 5), centered at y=430.
-const CANVAS_W = 640;
-const CANVAS_H = 480;
-const QUIT_ENTRY: [number, number] = [(400 + 179 / 2) / CANVAS_W, (20 + 5 * 76 + 30) / CANVAS_H];
+/** Quit is the last of the six main-menu entries. */
+const QUIT_ENTRY = menuEntry(5);
 
 test(
   "clicking Quit in the flat menu requests a quit without killing the runtime",
@@ -56,20 +54,6 @@ test(
 // The VR menu is the same canvas on a world panel, driven by a controller ray -
 // so the Quest's Quit has to travel the same path (AGENTS.md: "a canvas must
 // render the same way in flatscreen and in VR").
-//
-// The runtime's VR head yaw is +90 degrees (the camera looks along -X), so the
-// panel hangs at x=-2 facing the head: canvas +x maps to world -z and +y to -y.
-const PANEL_DISTANCE = 2;
-const PANEL_SIZE = { x: 2, y: 1.5 };
-/** 90-degree yaw: rotates a hand's -Z ray onto the panel's -X. */
-const AIM_AT_PANEL: [number, number, number, number] = [0, 0.7071068, 0, 0.7071068];
-
-/** Pawn-local position of a canvas point on the VR panel. */
-const panelPoint = ([u, v]: [number, number]): [number, number, number] => [
-  -PANEL_DISTANCE,
-  PLAYER_EYE_HEIGHT_WORLD + (0.5 - v) * PANEL_SIZE.y,
-  (0.5 - u) * PANEL_SIZE.x,
-];
 
 test(
   "the VR hand ray can quit from the menu too",
